requirements.txt
torch
transformers==4.11.3
easyocr==1.4.1
opencv-python==4.5.3.56
numpy==1.21.2
pandas==1.3.3
pytesseract==0.3.8
sort==2019.12.11
ultralytics==0.6.3
python-dotenv